Level1C 95% Complete
Enhance lane system further. Each wave of boats will now seem as if its really in the water. Boats will be assigned boat numbers. Boat numbers are based on which boat is on top of the hierarchy.  The boat number helps re assure that the boats are going to the right index of the panel as siblings. Now working on speedboat