Bug Fixes
Fix saving issue that will cause player to have to recreate a profile if they chose anything above pre-k in the tutorial level. Avatar now saves properly when selecting in tutorial level. Input field no longer allows any spaces. Will now check to see if there are any characters in the field when the player submits. If there are none, then the player will not be able to submit